Update review_normalizer.py. Add normalize_text function. Remove review_normalizer.py
