+++ src/review_cleaner.py
@@ -14,6 +14,49 @@ logger = setup_logger('review_cleaner.log')
 invalid_rating_count = 0
 
 
+def normalize_text(text: str) -> str:
+    """
+    Нормализует текст: удаляет лишние пробелы и знаки препинания, не примыкающие к слову.
+    :param text: Исходный текст.
+    :return: Нормализованный текст.
+    """
+    # Сначала удаляем лишние пробелы
+    text = re.sub(r"\s+", " ", text).strip()
+
+    # Удаляем пробелы перед знаками пунктуации
+    text = re.sub(r"\s([?.!,¿])", r"\1", text)
+    # Убираем лишние пробелы после знаков пунктуации
+    text = re.sub(r"([?.!,¿])\s*", r"\1 ", text)
+
+    # Удаляем пробелы между одинаковыми знаками пунктуации
+    text = re.sub(r"([?.!,¿])\s*\1+", r"\1", text)
+
+    # Разделяем текст на слова и восстанавливаем пробелы между ними
+    words = text.split()
+
+    # Удаляем знаки препинания, если они не являются частью слова
+    normalized_words = [
+        re.sub(r"[^\w\s,.!?()]+", "", word) for word in words
+    ]
+
+    # Собираем текст обратно из нормализованных слов
+    normalized_text = " ".join(normalized_words)
+
+    # Удаляем группы знаков пунктуации в конце текста
+    normalized_text = re.sub(r"[?.!,¿]+\s*$", "", normalized_text)
+
+    # Удаляем все подряд идущие знаки пунктуации в середине текста
+    normalized_text = re.sub(r"([?.!,¿])\1+", r"\1", normalized_text)
+
+    # Удаляем пробелы между знаками пунктуации и словами, если они есть
+    normalized_text = re.sub(r"\s([?.!,¿])", r"\1", normalized_text)
+
+    # Убираем пробел перед закрывающей скобкой
+    normalized_text = re.sub(r"\s\)", r")", normalized_text)
+
+    return normalized_text
+
+
 def clean_review_data(review: Dict[str, str]) -> Dict[str, str]:
     """
     Очищает данные отзыва, удаляя ненужные символы и проверяя корректность полей.
@@ -49,8 +92,8 @@ def clean_review_data(review: Dict[str, str]) -> Dict[str, str]:
         text = re.sub(r"[\n\r]+", " ", text)
         # Удаление одиночных символов 'n', которые могут оставаться после других преобразований
         text = re.sub(r"n", " ", text)
-        # Удаление лишних пробелов
-        text = re.sub(r"\s+", " ", text).strip()
+        # Применение нормализации текста
+        text = normalize_text(text)
         cleaned_review["text"] = text
 
         # Очистка рубрик (rubrics)
